Break base mailer into private and public bits
