+++ lib/templates/shiny/plugin/app/mailers/%namespaced_name%/application_mailer.rb.tt
@@ -2,9 +2,14 @@
 
 <%= wrap_in_modules <<~rb
   # Base mailer for #{camelized_modules}
-  class ApplicationMailer < ActionMailer::Base
-    default from: 'from@example.com'
-    layout 'mailer'
+  class ApplicationMailer < ShinyCMS::BaseMailer
+    def check_feature_flags
+      stop!
+    end
+
+    def check_ok_to_email
+      stop!
+    end
   end
 rb
 %>
