+++ plugins/ShinyAccess/app/mailers/shiny_access/application_mailer.rb
@@ -7,6 +7,5 @@
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
 module ShinyAccess
-  # Not used - Mailers inherit from ShinyCMS::ApplicationMailer
-  class ApplicationMailer < ActionMailer::Base; end
+  class ApplicationMailer < ShinyCMS::BaseMailer; end
 end
