+++ plugins/ShinyCMS/app/mailers/shinycms/application_mailer.rb
@@ -6,10 +6,6 @@
 #
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
-# Mailers that inherit from here must implement `check_feature_flags` (probably
-# using `enforce_feature_flags`) and `check_ok_to_email` (probably using either
-# `enforce_ok_to_email` or `enforce_do_not_contact`)
-
 module ShinyCMS
   # Base class for mailers - part of the ShinyCMS core plugin
   class ApplicationMailer < ActionMailer::Base
@@ -34,24 +30,6 @@ module ShinyCMS
       mail.perform_deliveries = false
       # Stop processing, now. See https://is.gd/ActionMailerStop (and last line of https://is.gd/ActionMailerCallbacks)
       self.response_body = nil
-    end
-
-    def enforce_feature_flags( *feature_flag_names )
-      stop! unless feature_flags_enabled? feature_flag_names
-    end
-
-    def enforce_ok_to_email( email_recipient )
-      stop! if email_recipient.blank?
-      # HasEmail.not_ok_to_email? checks on .confirmed status
-      # (AKA double opt-in) and DoNotContact status
-      stop! if email_recipient.not_ok_to_email?
-    end
-
-    def enforce_do_not_contact( email_address )
-      # You should use `enforce_ok_to_email` rather than this method whenever possible.
-      # The only exception is for mailer actions where checking the .confirmed status
-      # is impossible or would make no sense (e.g. when sending the confirmation email!)
-      stop! if DoNotContact.list_includes? email_address
     end
 
     def feature_flags_enabled?( *feature_flag_names )
