+++ plugins/ShinyForms/app/mailers/shiny_forms/application_mailer.rb
@@ -7,6 +7,5 @@
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
 module ShinyForms
-  # Not used - FormMailer inherits from ShinyCMS::ApplicationMailer
-  class ApplicationMailer < ActionMailer::Base; end
+  class ApplicationMailer < ShinyCMS::BaseMailer; end
 end
